--- NOTES.md
@@ -1,17 +1,0 @@
-## NOTES
-
-
-- How to determine maximum batch size?
-- What are the indications that training is going well or badly?
-- Is there any down side to my approach with validation script?
-- How would i incorporate some hdf5 saving into my workflow?
-
-#OVERFITTING
-- You will notice if training accuracy is much higher than validation
-
-
-#T1
-- After 13 epochs accuracy stops improving
-- Accuracy is very low (0.362)
-- After 2nd epoch validation accuracy freezes at 0.3360
-- during training the accuracy seemed much higher (~80) but the summaries tell a different story
