--- T1.logs
@@ -1,73 +1,0 @@
-988/1988 [==============================] - 156s 
-- loss: 1.0860 - acc: 0.4371 - val_loss: 1.1059 - val_acc: 0.2696                                                           
-Epoch 2/50                                                                                                                                                                    
-1988/1988 [==============================] - 145s 
-- loss: 1.1030 - acc: 0.3451 - val_loss: 1.1010 - val_acc: 0.3360                                                           
-Epoch 3/50                                                                                                                                                                    
-1988/1988 [==============================] - 145s 
-- loss: 1.1007 - acc: 0.3682 - val_loss: 1.0996 - val_acc: 0.3360                                                           
-Epoch 4/50                                                                                                                                                                    
-1988/1988 [==============================] - 146s 
-- loss: 1.0995 - acc: 0.3747 - val_loss: 1.0993 - val_acc: 0.3360                                                           
-Epoch 5/50                                                                                                                                                                    
-1988/1988 [==============================] - 145s 
-- loss: 1.1002 - acc: 0.3521 - val_loss: 1.0982 - val_acc: 0.3360                                                           
-Epoch 6/50                                                                                                                                                                    
-1988/1988 [==============================] - 147s 
-- loss: 1.0990 - acc: 0.3682 - val_loss: 1.0981 - val_acc: 0.3360                                                           
-Epoch 7/50                                                                                                                                                                    
-1988/1988 [==============================] - 149s 
-- loss: 1.0988 - acc: 0.3682 - val_loss: 1.0983 - val_acc: 0.3360                                                           
-Epoch 8/50                                                                                                                                                                    
-1988/1988 [==============================] - 148s 
-- loss: 1.1002 - acc: 0.3682 - val_loss: 1.0971 - val_acc: 0.3360                                                           
-Epoch 9/50                                                                                                                                                                    
-1988/1988 [==============================] - 147s 
-- loss: 1.0990 - acc: 0.3682 - val_loss: 1.0969 - val_acc: 0.3360                                                           
-Epoch 10/50                                                                                                                                                                   
-1988/1988 [==============================] - 147s 
-- loss: 1.0990 - acc: 0.3682 - val_loss: 1.0965 - val_acc: 0.3360                                                           
-Epoch 11/50                                                                                                                                                                   
-1988/1988 [==============================] - 124s 
-- loss: 1.0985 - acc: 0.3783 - val_loss: 1.0965 - val_acc: 0.3360                                                           
-Epoch 12/50                                                                                                                                                                   
-1988/1988 [==============================] - 124s 
-- loss: 1.0988 - acc: 0.3783 - val_loss: 1.0961 - val_acc: 0.3360                                                           
-Epoch 13/50                                                                                                                                                                   
-1988/1988 [==============================] - 121s 
-- loss: 1.0980 - acc: 0.3783 - val_loss: 1.0963 - val_acc: 0.3360                                                           
-Epoch 14/50                                                                                                                                                                   
-1988/1988 [==============================] - 119s 
-- loss: 1.0987 - acc: 0.3783 - val_loss: 1.0959 - val_acc: 0.3360                                                           
-Epoch 15/50                                                                                                                                                                   
-1988/1988 [==============================] - 120s 
-- loss: 1.0985 - acc: 0.3783 - val_loss: 1.0958 - val_acc: 0.3360                                                           
-Epoch 16/50                                                                                                                                                                   
-1988/1988 [==============================] - 121s 
-- loss: 1.0983 - acc: 0.3783 - val_loss: 1.0957 - val_acc: 0.3360    
-
-Epoch 17/50                                                                                                                                                                   
-1988/1988 [==============================] - 131s 
-- loss: 1.0983 - acc: 0.3783 - val_loss: 1.0955 - val_acc: 0.3360
-Epoch 18/50                                                                            
-1988/1988 [==============================] - 128s 
-- loss: 1.0977 - acc: 0.3783 - val_loss: 1.0957 - val_acc: 0.3360
-Epoch 19/50                                                                            
-1988/1988 [==============================] - 128s 
-- loss: 1.0983 - acc: 0.3783 - val_loss: 1.0956 - val_acc: 0.3360
-Epoch 20/50                                                                            
-1988/1988 [==============================] - 127s 
-- loss: 1.0983 - acc: 0.3783 - val_loss: 1.0955 - val_acc: 0.3360
-Epoch 21/50                                                                            
-1988/1988 [==============================] - 127s 
-- loss: 1.0981 - acc: 0.3783 - val_loss: 1.0954 - val_acc: 0.3360
-Epoch 22/50                                                                            
-1988/1988 [==============================] - 127s 
-- loss: 1.0981 - acc: 0.3783 - val_loss: 1.0954 - val_acc: 0.3360
-Epoch 23/50                                                                            
-1988/1988 [==============================] - 127s 
-- loss: 1.0979 - acc: 0.3783 - val_loss: 1.0955 - val_acc: 0.3360         
-Epoch 24/50                                
-1988/1988 [==============================] - 126s 
-- loss: 1.0981 - acc: 0.3783 - val_loss: 1.0954 - val_acc: 0.3360
-Epoch 25/50
