+++ cam.py
@@ -1,15 +1,31 @@
 from keras.models import *
 from keras.callbacks import *
+from keras.preprocessing import image
 from keras.preprocessing.image import ImageDataGenerator
 import keras.backend as K
 from model import *
 from data import *
-from utils import *
 import cv2
 import argparse
-from keras.utils.np_utils import to_categorical
 BATCH_SIZE = 32
-NB_EPOCHS = 50
+NB_EPOCHS = 40
+
+def get_batches(
+        dirname, 
+        gen=image.ImageDataGenerator(), 
+        shuffle=False, 
+        save_to_dir=None,
+        batch_size=32, 
+        class_mode='categorical',
+        target_size=(128,128)):
+    return gen.flow_from_directory(
+            dirname, 
+            save_to_dir=save_to_dir,
+            target_size=target_size,
+            class_mode=class_mode, 
+            shuffle=shuffle, 
+            batch_size=batch_size)
+
 
 def train(dataset_path):
         gen = ImageDataGenerator(
@@ -41,12 +57,11 @@ def train(dataset_path):
 
 def visualize_class_activation_map(model_path, img_path, output_path):
         model = load_model(model_path)
-        original_img = cv2.imread(img_path, 1)
+        original_img = cv2.resize(cv2.imread(img_path, 1), (224, 224)) 
         width, height, _ = original_img.shape
 
         #Reshape to the network input shape (3, w, h).
         img = np.array([np.transpose(np.float32(original_img), (2, 0, 1))])
-        
         #Get the 512 input weights to the softmax.
         class_weights = model.layers[-1].get_weights()[0]
         final_conv_layer = get_output_layer(model, "conv5_3")
@@ -54,16 +69,16 @@ def visualize_class_activation_map(model_path, img_path, output_path):
         [conv_outputs, predictions] = get_output([img])
         conv_outputs = conv_outputs[0, :, :, :]
 
-        print "predictions", predictions
-        target_class = 0
         #Create the class activation map.
         cam = np.zeros(dtype = np.float32, shape = conv_outputs.shape[1:3])
-        for i, w in enumerate(class_weights[:, target_class]):
+        for i, w in enumerate(class_weights[:, 1]):
                 cam += w * conv_outputs[i, :, :]
+        print "predictions", predictions
         cam /= np.max(cam)
         cam = cv2.resize(cam, (height, width))
         heatmap = cv2.applyColorMap(np.uint8(255*cam), cv2.COLORMAP_JET)
         heatmap[np.where(cam < 0.2)] = 0
+        print(heatmap.shape)
         img = heatmap*0.5 + original_img
         cv2.imwrite(output_path, img)
